converting the app to typescript
