+++ src/App.tsx
@@ -2,7 +2,6 @@ import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
-import './styles/index.css'
 import { BrowserRouter } from "react-router-dom";
 import MoviesContainer from './components/MoviesContainer'
